add telegram bot + refactor

+++ app/backend/db/models/user.py
@@ -1,4 +1,7 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger
+from sqlalchemy import Column
+from sqlalchemy import Integer
+from sqlalchemy import String
 
 from app.backend.db.base_model import Base
 from app.backend.db.mixins import TimeMixin
@@ -17,8 +20,11 @@ class ExternalUserModel(TimeMixin, Base):
     __tablename__ = "external_users"
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    telegram_id = Column(Integer, unique=True)
-    vk_id = Column(Integer, unique=True)
+    telegram_id = Column(
+        BigInteger,
+        unique=True,
+    )
+    vk_id = Column(BigInteger, unique=True)
     group_number = Column(Integer)
     first_name = Column(String)
     last_name = Column(String)
